requirements.txt
Flask==2.3.3
requests==2.31.0
openai==0.28.1
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
